html tag

+++ contracts/working/NFT.sol
@@ -8,9 +8,7 @@ import "@openzeppelin/contracts/token/ERC721/extensions/ERC721Enumerable.sol";
 
 import "@openzeppelin/contracts/access/Ownable.sol";
 
-
 import "hardhat/console.sol";
-
 
 contract NFT is ERC721URIStorage, ERC721Enumerable {
     using Strings for uint256;
@@ -18,29 +16,46 @@ contract NFT is ERC721URIStorage, ERC721Enumerable {
     Counters.Counter private _tokenIds;
     address contractAddress;
 
-    constructor() ERC721("NFT", "NFT") {
-    }
+    constructor() ERC721("NFT", "NFT") {}
 
     // Required overrides from parent contracts
-    function _burn(uint256 tokenId) internal virtual override(ERC721, ERC721URIStorage) {
+    function _burn(uint256 tokenId)
+        internal
+        virtual
+        override(ERC721, ERC721URIStorage)
+    {
         super._burn(tokenId);
     }
 
-    function tokenURI(uint256 tokenId) public view virtual override(ERC721, ERC721URIStorage) returns (string memory) {
+    function tokenURI(uint256 tokenId)
+        public
+        view
+        virtual
+        override(ERC721, ERC721URIStorage)
+        returns (string memory)
+    {
         // return string(abi.encodePacked(super.tokenURI(tokenId), ".json"));
         return super.tokenURI(tokenId);
     }
 
-    function _beforeTokenTransfer(address from, address to, uint256 tokenId) internal override(ERC721, ERC721Enumerable) {
+    function _beforeTokenTransfer(
+        address from,
+        address to,
+        uint256 tokenId
+    ) internal override(ERC721, ERC721Enumerable) {
         super._beforeTokenTransfer(from, to, tokenId);
     }
 
-    function supportsInterface(bytes4 interfaceId) public view override(ERC721, ERC721Enumerable) returns (bool) {
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        override(ERC721, ERC721Enumerable)
+        returns (bool)
+    {
         return super.supportsInterface(interfaceId);
     }
 
-    function createToken(string memory tokenURI) public returns (uint) {
-        
+    function createToken(string memory tokenURI) public returns (uint256) {
         uint256 newItemId = _tokenIds.current();
         _tokenIds.increment();
 
@@ -49,7 +64,4 @@ contract NFT is ERC721URIStorage, ERC721Enumerable {
 
         return newItemId;
     }
-
-
-
 }
